+++ composeApp/src/commonMain/kotlin/com/altmann/choresmanager/models/chores/college/CollegeChore.kt
@@ -22,7 +22,6 @@ class CollegeChore(
     priority: Priority,
     finishedDate: LocalDate,
     deleted: Boolean? = false,
-    userId: Int,
     finished: Boolean? = false,
     color: Color? = Color.Blue,
 
@@ -36,7 +35,7 @@ class CollegeChore(
 ) : Chore(
     choreId, startTime, endTime, daysOfWeek,
     startDate, endDate, choreException, title, description,
-    priority, finishedDate, deleted, userId, finished, color
+    priority, finishedDate, deleted, finished, color
 ) {
 
     init {